+++ bsp/my_adc.h
@@ -4,7 +4,7 @@
 #include "main.h"
 // extern float ext_vp;
 // extern uint8_t adc_value;
-uint8_t getADC(ADC_HandleTypeDef *hadc);
+float getADC(ADC_HandleTypeDef *hadc);
 
 
 #endif /*__MY_ADC_H__*/
